+++ cypress/locators/numeralDate/index.js
@@ -1,0 +1,6 @@
+import { NUMERAL_DATE_COMPONENT } from './locators';
+import { DATE_INPUT } from '../date-input/locators';
+
+export const numeralDateComponent = () => cy.iFrame(NUMERAL_DATE_COMPONENT);
+export const numeralDateInput = () => numeralDateComponent().find(DATE_INPUT);
+export const numeralDateInputByPosition = index => numeralDateInput().eq(index);