+++ cypress/locators/numeralDate/locators.js
@@ -1,0 +1,1 @@
+export const NUMERAL_DATE_COMPONENT = '[data-component="numeral-date"]';