+++ src/components/validations/validations.stories.js
@@ -63,7 +63,7 @@ const DeprecationWarning = () => (
     backgroundColor: 'red', textAlign: 'center', color: 'white', padding: 20, fontWeight: 'bold', marginBottom: 10
   } }
   >
-  Form and validations have been deprecated please see <a href='https://github.com/Sage/carbon/pull/2481'>#2481</a>
+    Form and validations have been deprecated please see <a href='https://github.com/Sage/carbon/pull/2481'>#2481</a>
   </div>
 );
 function makeBasicStory(name, themeSelector) {
@@ -164,7 +164,8 @@ function makeBasicStory(name, themeSelector) {
               </State>
             </Column>
           </Row>
-        </Form></>
+        </Form>
+      </>
     );
   };