+++ src/style/themes/base/base-theme.config.js
@@ -1,5 +1,6 @@
 export default (palette) => {
   const { atOpacity } = palette,
+
       blackWithOpacity = atOpacity('#000000'),
       baseWithOpacity = atOpacity(palette.productGreen);
 
@@ -11,7 +12,6 @@ export default (palette) => {
     colors: {
       // main
       base: palette.productGreen,
-
       primary: palette.genericGreenTint(15),
       secondary: palette.genericGreenShade(35),
       tertiary: palette.genericGreenShade(55),
@@ -42,7 +42,6 @@ export default (palette) => {
         hover: palette.errorRedShade(20)
       }
     },
-
 
     anchorNavigation: {
       divider: palette.slateTint(80),
@@ -78,6 +77,13 @@ export default (palette) => {
     carousel: {
       activeSelectorBackground: palette.slateTint(40),
       inactiveSelectorBackground: palette.slateTint(80)
+    },
+
+
+    flatTable: {
+      default: {
+        hover: palette.slateTint(95)
+      }
     },
 
     help: {
@@ -156,6 +162,11 @@ export default (palette) => {
 
     popoverContainer: {
       iconColor: 'rgba(0,0,0,0.90)'
+    },
+
+    numeralDate: {
+      passive: '#668592',
+      error: '#C7384F'
     },
 
     portrait: {